# Further data locality optimizations for implicit GEMM