## Pull Request Template

### Checklist

- [x] Confirmed that `run-checks all` script has been executed.
- [x] Made sure the book is up to date with changes in this PR.

### Related Issues/PRs

Optimizations for the implicit GEMM algorithm introduced in #2287 

### Changes

Optimizes the loop order and uses a transposed matrix for weights to optimize data access and allow vectorizing the input and weight.

### Testing

Ran the one test that currently fits the prerequisites for `implicit_gemm`

### Benchmarking
#### Before
| Benchmark | Feature  | Backend     | Device                  | Median         |
|-----------|----------|-------------|-------------------------|----------------|
| conv2d    | cuda-jit | `jit<cuda>` | CudaDevice { index: 0 } | 4.859ms        |
#### After
| Benchmark | Feature  | Backend     | Device                  | Median         |
|-----------|----------|-------------|-------------------------|----------------|
| conv2d    | cuda-jit | `jit<cuda>` | CudaDevice { index: 0 } | 2.792ms        |
| conv2d    | tch-gpu  | `tch`       | Cuda(0)                 | 1.888ms        |

`out_channels = 64`
| Benchmark | Feature  | Backend     | Device                  | Median         |
|-----------|----------|-------------|-------------------------|----------------|
| conv2d    | cuda-jit | `jit<cuda>` | CudaDevice { index: 0 } | 7.083ms        |
| conv2d    | tch-gpu  | `tch`       | Cuda(0)                 | 7.189ms        |