@laggui 2024-09-24
Just noticed the CI is failing due to the huge refactor we merged this morning (#2298). We wanted to merge your previous PRs before merging that one too so it wouldn't be annoying for you to have to deal with the changes lol.

We removed the const D generic from the primitives so some things changed slightly under the hood due to this (e.g., `permute` takes a slice instead of a fixed size array - as indicated by the failing CI).

It should be fairly straightforward to fix given the small code footprint of this PR 🙂 

But if you have any questions feel free to ping me here or on discord!